requirements.txt
feedparser>=6.0
requests>=2.28
beautifulsoup4>=4.12
google-api-python-client>=2.100
google-auth>=2.23
